Revert "Week3"
