+++ App.js
@@ -1,7 +1,8 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, TextInput, View } from 'react-native';
-import Header from './components/Header';
-import Input from './components/Input';
+import Header from './comonents/Header';
+import Input from './comonents/Input';
+import { useState } from 'react';
 
 export default function App() {
 
@@ -10,7 +11,7 @@ export default function App() {
   return (
     <View style={styles.container}>      
       <Header name={appName} />
-      <Input autoFocus={true}/>
+      <Input/>
       <StatusBar style="auto" />
     </View>
   );
